+++ modules/battlenetloop.py
@@ -2,10 +2,10 @@
 This module contains functions related to entering the game from Battle.net.
 """
 
+import logging
 import time
 from .constants import Button, Action
 from .image_utils import find_element
-import logging
 
 
 log = logging.getLogger(__name__)
@@ -15,11 +15,15 @@ def enter_from_battlenet():
     """
     Enters the game from the Battle.net application.
     """
-    find_element(Button.battlenet.filename, Action.move_and_click) and time.sleep(1)
+    found_element = find_element(Button.battlenet.filename, Action.move_and_click)
+    if found_element:
+        time.sleep(1)
 
-    find_element(
+    found_element = find_element(
         Button.battlenet_hearthstone.filename, Action.move_and_click
-    ) and time.sleep(1)
+    )
+    if found_element:
+        time.sleep(1)
 
     if find_element(Button.battlenet_play.filename, Action.move_and_click):
         log.info("Waiting (20s) for game to start")